refactor(LottoMachine.java) : ManualLottoManager 제거하여 생성 메소드 통일
